renamed CsvGenerator to CsvGeneratorClass
